RAG_BOT/requirements.txt
# Core Constraints (pypdf and chromadb are sensitive to numpy)
numpy<2.0.0
protobuf==3.20.0

# Core API
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic==2.6.1
requests==2.31.0

# LangChain Ecosystem (Pinned for compatibility)
langchain==0.3.0
langchain-community==0.3.0
langchain-core==0.3.0
langchain-text-splitters==0.3.0
langchain-chroma==0.1.4
langchain-google-genai==2.0.0

# Vector DB & AI
chromadb==0.5.5
sentence-transformers==3.0.1
pypdf==4.2.0

# Auth & Database
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
sqlalchemy==2.0.27
importlib-metadata>=6.0.0
pymysql==1.1.0
python-dotenv==1.0.1

